backend/requirements.txt
fastapi
uvicorn
requests
python-jose[cryptography]
python-dotenv
